Allow output configuration
Adds `--verbose` and `--quiet` options to the command line program along
with changing the majority of `console.log` calls as debug messages.
Output is now done through a log instance, which is a singleton,
configured once at the start of the command and usable elsewhere.

Future modules can use current Logger only by requiring it, as it is a
Singleton, configuration will already be present and only relevant log
levels will fo to the output or configured log file.